requirements.txt
requests
beautifulsoup4
httpx
python-dotenv
google-genai
